[FIX] naming InterviewerSearchField to SearchField && [ADD] key down event(Arrow Down, Arrow Up, Enter)
